refactor, 3 new src files for photo, video and channel
